requirements.txt
# Azure ML and Cloud
azure-ai-ml==1.12.0
azure-identity==1.15.0
azureml-core==1.54.0
azureml-mlflow==1.54.0

# ML and Deep Learning
torch==2.1.0
torchvision==0.16.0
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
pillow==10.1.0

# API and Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Data handling
kagglehub==0.2.5
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
mlflow==2.8.1

# Testing
pytest==7.4.3
httpx==0.25.2
